+++ src/main/java/com/tavisca/trainings/models/Suggestion.java
@@ -1,4 +1,4 @@
-package com.tavisca.trainings.services;
+package com.tavisca.trainings.models;
 
 public class Suggestion {
 	private String content;
