+++ src/main/java/com/tavisca/trainings/services/Handler.java
@@ -4,6 +4,9 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import com.tavisca.trainings.models.Suggestion;
+import com.tavisca.trainings.models.Todo;
+
 public class Handler {
 
 	List<Todo> todoList;
